Separate CodeMirror from FilesExplorer
